Fix precision warnings

+++ Sub-Terra/include/OpenSimplexNoise.h
@@ -4,12 +4,12 @@ class OpenSimplexNoise {
 public:
 	typedef float DecimalType;
 private:
-	const DecimalType STRETCH_CONSTANT_2D = -0.211324865405187;   // (1/Math.sqrt(2+1)-1)/2;
-	const DecimalType SQUISH_CONSTANT_2D  =  0.366025403784439;   //   (Math.sqrt(2+1)-1)/2;
-	const DecimalType STRETCH_CONSTANT_3D = -1.0 / 6;             // (1/Math.sqrt(3+1)-1)/3;
-	const DecimalType SQUISH_CONSTANT_3D  =  1.0 / 3;             //   (Math.sqrt(3+1)-1)/3;
-	const DecimalType STRETCH_CONSTANT_4D = -0.138196601125011;   // (1/Math.sqrt(4+1)-1)/4;
-	const DecimalType SQUISH_CONSTANT_4D  =  0.309016994374947;   //   (Math.sqrt(4+1)-1)/4;
+	const DecimalType STRETCH_CONSTANT_2D = static_cast<DecimalType>(-0.211324865405187);   // (1/Math.sqrt(2+1)-1)/2;
+	const DecimalType SQUISH_CONSTANT_2D  = static_cast<DecimalType>( 0.366025403784439);   //   (Math.sqrt(2+1)-1)/2;
+	const DecimalType STRETCH_CONSTANT_3D = static_cast<DecimalType>(-1.0 / 6);             // (1/Math.sqrt(3+1)-1)/3;
+	const DecimalType SQUISH_CONSTANT_3D  = static_cast<DecimalType>( 1.0 / 3);             //   (Math.sqrt(3+1)-1)/3;
+	const DecimalType STRETCH_CONSTANT_4D = static_cast<DecimalType>(-0.138196601125011);   // (1/Math.sqrt(4+1)-1)/4;
+	const DecimalType SQUISH_CONSTANT_4D  = static_cast<DecimalType>( 0.309016994374947);   //   (Math.sqrt(4+1)-1)/4;
 
 	const DecimalType NORM_CONSTANT_2D = 47;
 	const DecimalType NORM_CONSTANT_3D = 103;
